refactor(tests): split endpoint-routing tests out of test_models.py (CodeScene gate)
CodeScene flagged Low Cohesion (critical rule) after test_models.py
grew to five loosely-related classes. Move TestRuntimeBaseSplit and
TestRehostToContainer into test_endpoint_routing.py — one file for the
model listing/client surface, one for endpoint construction and the
ENG-8766 re-host rule. Pure move, no test changes.
